Add PageNav to location
