+++ src/App.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { BrowserRouter as Router, Switch, Route } from 'react-router-dom'
 import Navigation from './components/Navigation'
-import LocationPage from './components/pages/LocationPage'
+import LocationPage from './components/pages/Locations/LocationPage'
 import EpisodesPage from './components/pages/Episodes/EpisodesPage'
 import CharactersPage from './components/pages/CharactersPage'
 import styled from 'styled-components'
